requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
boto3==1.9.174
botocore==1.12.174
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.2
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
h2==3.1.0
hiredis==1.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
msgpack==0.6.1
Pillow==6.0.0
priority==1.3.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.1
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
urllib3==1.25.3
whitenoise==4.1.2
zope.interface==4.6.0
